requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
pyodbc==5.2.0
firebase-admin==6.9.0
python-dotenv==1.1.1
azure-monitor-opentelemetry==1.6.10
opentelemetry-instrumentation-fastapi==0.49b2